🛡️ Guardian (JULES): Consolidate formatting utilities
WHAT:
- Merged `src/components/ide/utils.ts` into `src/lib/format.ts`
- Updated import in `src/components/ide/prompt-ide.tsx`
- Deleted `src/components/ide/utils.ts`
- Added tests for `toYaml` in `src/__tests__/lib/format.test.ts`
- Created JULES tracking directory and files (`.Jules/JULES.md`, `.Jules/task-log.md`, `.Jules/guardian/...`)

WHY:
- Consolidates formatting utilities into a single module to prevent architectural scattering and duplicate implementations.

JULES CONTEXT:
- Checked for Autonomous conflicts.
- Session logged to `.Jules/guardian/2026-04-09/`.

VERIFICATION:
- Tests pass (`src/__tests__/lib/format.test.ts` extended and successful).
- Linter clean of newly introduced errors.
