+++ .Jules/JULES.md
@@ -1,0 +1,1 @@
+# JULES Standard
