+++ .Jules/task-log.md
@@ -1,0 +1,14 @@
+## 2026-04-09 16:36 - [GUARDIAN] Session started
+- Directory: .Jules/guardian/2026-04-09
+- Phase: PRE-FLIGHT
+- JULES Check: COMPLETE
+
+## 2026-04-09 17:08 - [GUARDIAN] Verification complete
+- Tests: ✓
+- Types: ✓
+- Lint: ✓
+- Build: ✓
+## 2026-04-09 17:08 - [GUARDIAN] Consolidated formatting utilities
+- Merged `src/components/ide/utils.ts` into `src/lib/format.ts`
+- Updated import in `src/components/ide/prompt-ide.tsx`
+- Deleted `src/components/ide/utils.ts`
